backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg[binary]>=3.1.18
python-dotenv==1.0.1
PyJWT==2.8.0
xgboost==2.0.3
pandas==2.2.2
scikit-learn==1.4.2
supabase==2.4.5
